move footer to component
